Added the HandwrittenClientTrait and removed references to ClientTrait in PubSubClient
